+++ module/dm.code.module.pipeline/src/pipeline/feature/xasset.cpp
@@ -13,7 +13,7 @@ namespace dm
 		HashBuilder hb;
 		hb.add_type<XFeatureAssetObject>();
 		hb.add(XASSET_CODE_VERSION_FEATURE);
-		hb.add(sig.symbol);
+		hb.add(static_cast<u64>(sig.symbol));
 		return hb.get();
 	}
